requirements.txt
# CPU-friendly, Docker-buildable versions
torch==2.4.1
torchvision==0.19.1
timm==1.0.20

# API server
fastapi==0.115.4
uvicorn[standard]==0.30.6
gunicorn==21.2.0

# Imaging
pillow==10.4.0